% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootspatreg.R
\name{ApproxSERoot2}
\alias{ApproxSERoot2}
\title{Linijka nr 1 - function title}
\usage{
ApproxSERoot2(model_spatial)
}
\arguments{
\item{model_spatial}{Spatial model object to analyse}
}
\value{
\code{ApproxSERoot2()} returns ... to be done.
}
\description{
Linijka nr 2 - description
}
\details{
Linijka nr 3 - details
}
\examples{
#To be done!!!

}
