% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_bootdbscan-package.R
\docType{package}
\name{bootdbscan-package}
\alias{bootdbscan-package}
\title{bootdbscan: To be done - package title}
\description{
TO BE DONE (package description) - More about what it does (maybe more than one line)
Use four spaces when indenting paragraphs within the Description.
}
\references{
To be done
}
\author{
Katarzyna Kopczewska and Mateusz Kopyt
}
