% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootspatreg.R
\name{BootSpatReg}
\alias{BootSpatReg}
\title{Spatial bootstrapped regression model for handling large data sets}
\usage{
BootSpatReg(points_sf, iter, sample_size, eq, model_type, knn)
}
\arguments{
\item{points_sf}{Geo-located points in \code{sf} or the \code{data.frame} class - in the case of a \code{data.frame} object, the first and second columns must contain X and Y coordinates.}

\item{iter}{The number of iterations.}

\item{sample_size}{The sample size, must be less than or equal to the number of points in the dataset (\code{points_sf} parameter). If \code{sample_size} is greater, it is automatically set
to the number of points in the dataset.For the first trial, try a value around 1000 for computational efficiency.}

\item{eq}{An object that defines the equation for the model, can be in the \code{\link[stats:formula]{stats::formula()}} class (or one that can be coerced to that class) or
a symbolic description of the model to be used.}

\item{model_type}{The type of spatial econometric model, one of "SAR", "SDM", "SEM". \emph{Soon to be updated for use with "SDEM", "SAC" models.}}

\item{knn}{The number of k nearest neighbours (knn) used to construct the spatial weight matrix based on the k nearest neighbours criterion.}
}
\value{
\code{BootSpatReg()} returns the following list object:
\item{coef.boot}{A \code{data.frame} with the coefficients of all iterations.}
\item{error.boot}{A \code{data.frame} with the standard errors of the coefficients from all iterations.}
\item{quality.boot}{A set of quality metrics for each iteration: Akaike Information Criterion of OLS model (\code{AIC.ols}) and selected spatial model (\code{AIC.spatial}),
time of computation of spatial weight matrix (\code{time.W}) and estimation of spatial model (\code{time.model}), selected spatial coefficient, rho or lambda (\code{spatial.coef}).}
\item{data.best}{A subset of the data used to estimate the best model, reported as an \code{sf} object; the rownames (\code{outome$data.best}) command allows you to obtain the IDs of the rows
used in the iteration that produced the best model.}
\item{knnW.best}{A spatial weight matrix used to estimate the best spatial model, constructed using the k nearest neighbours criterion for k specified by the user as input.}
\item{model.best}{An object of the best spatial model, selected with PAM from the set of bootstrap models.}
\item{RAMSE.best}{Root Mean Square Error (RAMSE) of the best spatial model.}
}
\description{
The function uses the bootstrap approach to estimate a set of spatial models, each on a different subset of the data - the function selects the rows
and the columns are fixed according to the equation. The best model is selected using the Partitioning Around Medoid (PAM) algorithm for the k=1 cluster -
it is the most central model in a multidimensional setting. The function reports which observations are associated with the best model.
}
\details{
For each iteration specified in the iter argument, the function selects the random subset of observations, estimates the model and stores the results.
The best model is selected as the most central using multidimensional medoid for all estimated coefficients. PAM searches for the model that is closest
to the other models on all coefficients.

The models are estimated more quickly on subsamples than on a full sample. As shown in Kopczewska (2023), coefficients in subsample models are consistent -
there is no significant difference between coefficients from full and subsample models. The standard errors of the coefficients in a subsample model
are generally higher than in a full sample model. However, they can be approximated by the √2 rule - the standard error in a model based on a sample twice as large
is √2 as small. This approximation can be easily checked using the \code{\link[=ApproxSERoot2]{ApproxSERoot2()}} function. This property allows estimating spatial econometric models
for large data using random subsamples without losing precision.

\code{BootSpatReg()} is a time-efficient, methodologically correct approach to deal with the non-scalability of the spatial weight matrix W.
}
\examples{
# The form of the equation to be estimated:
# Companies' return on assets (ROA) depends on their size, sector and relative location.
eq<-roa~empl+dummy.prod+dummy.constr+dummy.serv+dist.big.city

# Bootstrapped model estimation - it can take a long while.
bsr<-BootSpatReg(firms_sf, 5, 1500, eq, "SEM", knn=5)

# Result display
bsr					# full output
head(rownames(bsr$data.best),10)	# IDs of data used in the best model
summary(bsr$model.best)	# summary with p-value of the best model

# for this ouput use ApproxSERoot2() to approximate the standard errors for bigger sample
# or use SpatPredTess() to predict for out-of-sample points

}
\references{
Kopczewska, K. (2023). Spatial bootstrapped microeconometrics: Forecasting for out‐of‐sample geo‐locations in big data.
Scandinavian Journal of Statistics.
}
