% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta.R
\name{ETA}
\alias{ETA}
\alias{eta}
\title{Linijka nr 1 - function title}
\usage{
ETA(points_sf, region_sf, sample_size)

eta(points_sf, region_sf, sample_size)
}
\arguments{
\item{points_sf}{do opisu (obiekt sf lub data.frame - w data frame 1 kolumna musi być X coords, druga kolumna Y coords). When using a simple data.frame, make sure that the coordinates of the points are in the same coordinate system / projection as the \code{region_sf} object.}

\item{region_sf}{do opisu (obiekt sf ale jako region)}

\item{sample_size}{Sample size, must be less than or equal to the number of points in the dataset (\code{points_sf} parameter). If \code{sample_size} is larger, it is automatically set to the number of points in the dataset. We suggest that a value greater than 800 is not used for reasons of computational efficiency. (SPRAWDZIĆ)}
}
\value{
\code{ETA()} returns ... to be done.
}
\description{
Linijka nr 2 - description
}
\details{
Linijka nr 3 - details
}
\examples{
#To be done!!!

}
