% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{firms_sf}
\alias{firms_sf}
\title{A collection of data on companies in the Lubelskie Voivodship in Poland.}
\format{
An object of class \code{sf} with projection EPSG:4326 (WGS84) consisting of the following variables:
\describe{
\item{empl}{Employment.}
\item{roa}{Return on assets.}
\item{dummy.agri}{Dummy variable indicating whether the enterprise belongs to the agricultural sector.}
\item{dummy.prod}{Dummy variable indicating whether the enterprise belongs to the production sector.}
\item{dummy.constr}{Dummy variable indicating whether the enterprise belongs to the construction sector.}
\item{dummy.serv}{Dummy variable indicating whether the enterprise belongs to the services sector.}
\item{dist.big.city}{Distance (in km) to the nearest large city.}
\item{dist.closest.city}{Distance (in km) to the nearest city.}
\item{year}{Year.}
\item{sector.aggr}{Sector to which the enterprise belongs.}
\item{sector}{Sector according to Polish Classification of Economic Activities (PKD 2007).}
}
}
\source{
Own work based on REGON registry (Statistics Poland),
\url{https://wyszukiwarkaregon.stat.gov.pl/}
}
\usage{
firms_sf
}
\description{
A collection of data on enterprises in the Lubelskie Voivodeship in Poland, based on the REGON register provided by the Polish Statistical Office
(Statistics Poland - Główny Urząd Statystyczny, GUS). It consists of a 10\% representative sample of enterprises registered in 2012 in the NUTS2 region
of Lubelskie.
}
\examples{
#To be done!!!
}
\keyword{datasets}
