% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdc.R
\name{QDC}
\alias{QDC}
\alias{qdc}
\title{Linijka nr 1 - function title}
\usage{
QDC(data_sf, sample_size, nclust = 3, k = 10, eps = 0.05)
}
\arguments{
\item{data_sf}{do opisu (obiekt sf lub data.frame - w data frame 1 kolumna musi być X coords, druga kolumna Y coords)}

\item{sample_size}{Sample size, must be less than or equal to the number of points in the dataset (\code{data_sf} parameter). If \code{sample_size} is larger, it is automatically set to the number of points in the dataset.}

\item{nclust}{Number of clusters to find. Must be >=2. For nclust=3 additional cluster labels will be included in the output object.}

\item{k}{Number of nearest neighbors used for the distance calculation. Default value set to 10.}

\item{eps}{do opisu}
}
\value{
\code{QDC()} returns ... to be done.
}
\description{
Linijka nr 2 - description
}
\details{
Linijka nr 3 - details
}
\examples{
#To be done!!!

}
