% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{region_sf}
\alias{region_sf}
\title{Boundaries of the Lubelskie Voivodship in Poland.}
\format{
An object of class \code{sf} with projection EPSG:4326 (WGS84) consisting of the boundaries of the Lubelskie Voivodeship in Poland.
}
\source{
Based on data provided by Head Office of Geodesy and Cartography (GUIK),
\url{https://www.geoportal.gov.pl/pl/dane/panstwowy-rejestr-granic-prg/}
}
\usage{
region_sf
}
\description{
Boundaries of the Lubelskie Voivodship in Poland (NUTS2 region) in object class \code{sf} obtained on the basis of shapefiles
of the Polish Boundary Register (PRG) provided by Head Office of Geodesy and Cartography in Poland
(Głowny Urząd Geodezji i Kartografii, GUGIK)
}
\examples{
#To be done!!!
}
\keyword{datasets}
