% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spag.R
\name{SPAG}
\alias{SPAG}
\alias{spag}
\title{Linijka nr 1 - function title}
\usage{
SPAG(points_sf, size_var, region_sf)

spag(points_sf, size_var, region_sf)
}
\arguments{
\item{points_sf}{do opisu (obiekt sf lub data.frame - w data frame 1 kolumna musi być X coords, druga kolumna Y coords). When using a simple data.frame, make sure that the coordinates of the points are in the same coordinate system / projection as the \code{region_sf} object. (OD RAZU DANE Z WYBRANEJ ZMIENNEJ NP. SEKTORA)}

\item{size_var}{The name of the variable describing the size of the analysed objects (e.g. companies). (SPRAWDZIĆ opis)}

\item{region_sf}{do opisu (obiekt sf ale jako region)}
}
\value{
\code{SPAG()} returns ... to be done.
}
\description{
Linijka nr 2 - description
}
\details{
Linijka nr 3 - details
}
\examples{
#To be done!!!

}
