% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benford.R
\name{SpatBenfordPattern}
\alias{SpatBenfordPattern}
\alias{spatbenfordpattern}
\title{Linijka nr 1 - function title}
\usage{
SpatBenfordPattern(region_sf, sample_size = 5000)

spatbenfordpattern(region_sf, sample_size = 5000)
}
\arguments{
\item{region_sf}{do opisu (obiekt sf ale jako region)}

\item{sample_size}{Number of points to be generated based on the region. Default value 5000. We suggest that a value greater than ..... is not used for reasons of computational efficiency. (SPRAWDZIĆ, UWAGA! st_sample czasem generuje mniej punktów)}
}
\value{
\code{SpatBenfordPattern()} returns ... to be done.
}
\description{
Linijka nr 2 - description
}
\details{
Linijka nr 3 - details
}
\examples{
#To be done!!!

}
\references{
#To be done - ważne tutaj bo sztywne parametry w kodzie
}
