% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_spatialWarsaw-package.R
\docType{package}
\name{spatialWarsaw-package}
\alias{spatialWarsaw}
\alias{spatialWarsaw-package}
\title{spatialWarsaw: Spatial analysis on geolocalised point data including clustering measures and bootstrap methods}
\description{
A collection of tools for spatial analysis on geolocalised point data including clustering measures and the ability to bootstrap large datasets. The functions in this package use spatial machine learning, spatial statistics and spatial econometric approaches.
}
\references{
To be done
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poktam/spatialWarsaw}
  \item Report bugs at \url{https://github.com/poktam/spatialWarsaw/issues}
}

}
\author{
\strong{Maintainer}: Mateusz Kopyt \email{mkopyt@wne.uw.edu.pl} (\href{https://orcid.org/0000-0001-5285-7792}{ORCID})

Authors:
\itemize{
  \item Katarzyna Kopczewska \email{kkopczewska@wne.uw.edu.pl} (\href{https://orcid.org/0000-0003-1065-1790}{ORCID})
  \item Maria Kubara \email{maria.kubara@uw.edu.pl} (\href{https://orcid.org/0000-0002-8768-8391}{ORCID})
}

Other contributors:
\itemize{
  \item Ewa Dobrowolska [contributor]
}

}
\keyword{internal}
