% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootspatreg.R
\name{SpatPredTess}
\alias{SpatPredTess}
\title{Linijka nr 1 - function title}
\usage{
SpatPredTess(model_spatial, points_spatial_sf, knnW, points_new_sf, region_sf)
}
\arguments{
\item{model_spatial}{Spatial model}

\item{points_spatial_sf}{Data on which the model_spatial was estimated (obiekt sf lub data.frame - w data frame 1 kolumna musi być X coords, druga kolumna Y coords).
When using a simple data.frame, make sure that the coordinates of the points are in the same coordinate system / projection as the \code{region_sf} object.}

\item{knnW}{Spatial weighting matrix (\code{listw} class) used to estimate model_spatial}

\item{points_new_sf}{New data for prediction (all used for a forecast). NOTE: The new data must have the same class and structure as
the \code{points_spatial_sf} object. (obiekt sf lub data.frame - w data frame 1 kolumna musi być X coords, druga kolumna Y coords).
When using a simple data.frame, make sure that the coordinates of the points are in the same coordinate system / projection as the \code{region_sf} and \code{points_spatial_sf} objects.}

\item{region_sf}{do opisu (obiekt sf ale jako region)}
}
\value{
\code{SpatPredTess()} returns ... to be done.
}
\description{
Linijka nr 2 - description
}
\details{
Linijka nr 3 - details
}
\examples{
#To be done!!!

}
