% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssr.R
\name{ssr}
\alias{ssr}
\title{ssr() function}
\usage{
ssr(
  data_sf,
  type = "ClustConti",
  clusters,
  sep,
  r_p = 0.001,
  eps_r = 1e-15,
  eps_np = 0.01,
  minPts = 5,
  bootstrap = FALSE,
  sample_size = NULL,
  times = NULL,
  eq,
  family = binomial,
  ...
)
}
\arguments{
\item{data_sf}{Geo-located points in \code{sf} or the \code{data.frame} class - in the case of a \code{data.frame} object, the first and second columns must contain X and Y coordinates.}

\item{type}{One of the: "ClustConti","ClustDisjoint". See \code{\link[=ClustConti]{ClustConti()}} and \code{\link[=ClustDisjoint]{ClustDisjoint()}} for more details.}

\item{clusters}{do opisu}

\item{sep}{do opisu}

\item{r_p}{do opisu}

\item{eps_r}{do opisu}

\item{eps_np}{do opisu}

\item{minPts}{do opisu}

\item{bootstrap}{do opisu}

\item{sample_size}{do opisu (used only if bootstrap=TRUE)}

\item{times}{do opisu  (used only if bootstrap=TRUE)}

\item{eq}{an object of class \code{\link[stats:formula]{stats::formula()}} (or one that can be coerced to that class):
a symbolic description of the model to be used.}

\item{family}{A \code{family} object  consistent with \code{\link[stats:family]{stats::family()}}: a description of the error distribution
and link function to be used in the model. The default value is \code{binomial} with logit link function.}

\item{...}{Other parameters passed to \code{\link[stats:glm]{stats::glm()}} during the calculation.}
}
\value{
Will be updated soon
}
\description{
Will be updated soon
}
\details{
Will be updated soon
}
\examples{
#To be done!!!

}
