% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssr.R
\name{ssrShowModels}
\alias{ssrShowModels}
\title{Friendly summary and plot for ssr objects (ssr() function result)}
\usage{
ssrShowModels(ssr_model, plot = TRUE, ...)
}
\arguments{
\item{ssr_model}{Result object of the \code{\link[=ssr]{ssr()}} function from this package. The class of \code{ssr_model} object must only be 'ssr'.}

\item{plot}{Logical; indicates whether the function should generate a plot, default \code{plot=TRUE}.}

\item{...}{Other parameters passed to \code{\link[stargazer:stargazer]{stargazer::stargazer()}}.}
}
\value{
This function returns the summary table of econometric models in density clusters. If \code{plot=TRUE} it will also return
the the point plot with division into clusters.
}
\description{
The \code{ssrShowModels()} function with the \code{plot=TRUE} option shows the summary of the models and plots the points divided
by the applied density clusters.
}
\details{
The plot illustrates the division of geolocated points into density clusters. This partitioning is used in
the \code{\link[=ssr]{ssr()}} function for subgroup model estimation.
}
\examples{
# The following example is on selected data from the firms_sf dataset included in the package
sub_sf<-firms_sf[1:5000,]
# One to choose
# clust<-ClustConti(sub_sf, clusters=4, noise =0.2, minPts0=30)
clust<-ClustDisjoint(sub_sf, noise=c(0.8, 0.6, 0.4, 0.2), minPts=30)

# Density cluster plot
sub_sf$cluster<-clust$cluster
plot(sub_sf["cluster"], key.pos=4, pch=".", cex=2, main="Density clusters")

# SSR models for the binary dependent variable
# The form of the equation to be estimated:
eq<-dummy.prod~empl+roa+dist.big.city

my.ssr<-ssr(sub_sf, clust, eq, family=binomial)
my.ssr

# Output with plot using ssrShowModels() function.
ssrShowModels(my.ssr, plot=TRUE)

}
