% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastclust.R
\name{STS}
\alias{STS}
\alias{sts}
\title{Linijka nr 1 - function title}
\usage{
STS(
  points_sf,
  eq,
  time_var,
  region_sf,
  nrows.raster = 50,
  ncols.raster = 50,
  nc,
  bw,
  adaptive = FALSE
)
}
\arguments{
\item{points_sf}{do opisu (obiekt sf lub data.frame - w data frame 1 kolumna musi być X coords, druga kolumna Y coords).
When using a simple data.frame, make sure that the coordinates of the points are in the same
coordinate system / projection as the \code{region_sf} object. NOTE! Data must be from a single period!}

\item{eq}{an object of class \code{\link[stats:formula]{stats::formula()}} (or one that can be coerced to that class):
a symbolic description of the model to be used.}

\item{time_var}{Name of the variable categorising the observations over time in the points_sf object.}

\item{region_sf}{do opisu (obiekt sf ale jako region)}

\item{nrows.raster}{raster row dimensions, default 50}

\item{ncols.raster}{raster column dimensions, default 50}

\item{nc}{number of clusters}

\item{bw}{do opisU: !!! ZOSTAWIŁEM bw W PARAMETRACH JAK W POPRZEDNIEJ FUNKCJI - JAK NIE BĘDZIE, TO BĘDZIEMY LICZYĆ DLA I OKRESU}

\item{adaptive}{is adaptive (do opisu), default=FALSE}
}
\value{
\code{STS()} returns ... to be done.
}
\description{
Linijka nr 2 - description
}
\details{
Linijka nr 3 - details
}
\examples{
#To be done!!!

}
