% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{testA}
\alias{testA}
\title{A TU BĘDZIE CAŁA POMOC DO FUNKCJI testA - chyba, że je ukryjemy i nie będą dostępne z zewnątrz ;)}
\usage{
testA(
  crds,
  clusters,
  sep,
  r_p = 0.001,
  eps_r = 1e-15,
  eps_np = 0.01,
  minPts0 = 16
)
}
\description{
A TU BĘDZIE CAŁA POMOC DO FUNKCJI testA - chyba, że je ukryjemy i nie będą dostępne z zewnątrz ;)
}
