% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{testB}
\alias{testB}
\title{A TU BĘDZIE CAŁA POMOC DO FUNKCJI testB - chyba, że je ukryjemy i nie będą dostępne z zewnątrz ;)}
\usage{
testB(
  crds,
  minPts = 5,
  sep = c(0.8, 0.6, 0.4, 0.2),
  r_p = 0.001,
  eps_r = 1e-15,
  eps_np = 0.01,
  bootstrap = FALSE,
  sample_size,
  times
)
}
\description{
A TU BĘDZIE CAŁA POMOC DO FUNKCJI testB - chyba, że je ukryjemy i nie będą dostępne z zewnątrz ;)
}
