% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootspatreg.R
\name{ApproxSERoot2}
\alias{ApproxSERoot2}
\title{Approximation of standard errors of coefficients from small to large sample using the √2 rule}
\usage{
ApproxSERoot2(model_spatial)
}
\arguments{
\item{model_spatial}{Spatial model object to analyse.}
}
\value{
\code{ApproxSERoot2()} returns a matrix of approximate standard errors for given variables and dataset sizes, and a linear graph
to graphically illustrate these relationships.
}
\description{
The function approximates the standard errors of the coefficients of any econometric model on n observations to hypothetical values
for other sample sizes: \code{n*2}, \code{n*4}, \code{n*8}, \code{n*16}, \code{n*32} observations.
}
\details{
Approximating the standard errors of the coefficients is important when using bootstrap regression, e.g. \code{\link[=BootSpatReg]{BootSpatReg()}}, which by default
estimates the model on a smaller dataset and by nature yields higher standard errors of the coefficients. According to Kopczewska (2023),
standard errors of coefficients, especially in spatial econometric models, follow the √2 rule - the standard error in a model on a sample
twice as large is √2 as small.

This function applies to hypothetical dataset sizes that are doubled \code{n*2}, \code{n*4}, \code{n*8}, \code{n*16}, \code{n*32}. Errors for these dataset sizes
are calculated as: \code{er/(2^0.5)^1}, \code{er/(2^0.5)^2}, \code{er/(2^0.5)^3}, \code{er/(2^0.5)^4}, \code{er/(2^0.5)^5}, where er is the standard error
from an original model for n observations.
}
\examples{
#To be done!!!

}
\references{
Kopczewska, K. (2023). Spatial bootstrapped microeconometrics: Forecasting for out‐of‐sample geo‐locations in big data. Scandinavian Journal of Statistics.
}
