% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/W_matrix.R
\name{bestW}
\alias{bestW}
\title{Best number of k nearest neighbours (knn) to construct a spatial weight matrix for point data}
\usage{
bestW(points_sf, eq, model_type = "SDM", sample_size, knn)
}
\arguments{
\item{points_sf}{Geo-located points in \code{sf} or the \code{data.frame} class - in the case of a \code{data.frame} object, the first and second columns must contain X and Y coordinates.}

\item{eq}{An object of class \code{\link[stats:formula]{stats::formula()}} (or one that can be coerced to that class) that defines the equation for the model: a symbolic description of the model to be used.}

\item{model_type}{The type of spatial econometric model, one of "SAR", "SDM", "SEM", "SDEM", "SAC". The default is "SDM".}

\item{sample_size}{The sample size, must be less than or equal to the number of points in the dataset (\code{points_sf} parameter). If \code{sample_size} is greater, it is automatically set
to the number of points in the dataset.}

\item{knn}{The vector of alternative k nearest neighbours used in the subsequent spatial econometric models for the spatial weight matrix.}
}
\value{
\code{bestW()} returns a spatial weight matrix that minimises the AIC of the models for the given equation.
It defines the best number of knn. It displays two line plots with AIC and spatial parameter (rho or lambda),
both depending on the user-defined knn.
\itemize{
\item \strong{Will be available soon} *: In addition, if the value of sample_size is less than the number of observations in the dataset, the function will return which observations were used in the modelling.
}
}
\description{
Function calculates a set of candidate spatial econometric models on point data. The K nearest neighbours included in the
spatial weight matrix are specified in a user-specified vector. For the same equation, the function calculates models using
different spatial weight matrices and compares the AIC (Akaike Information Criterion). The model with the lowest AIC
is selected as the best. The function outputs the spatial weight matrix W for the best model and the number of knn used in the best W.
}
\details{
Function in an iterative quality check of spatial models that differ only in the number of nearest neighbours used to construct
the spatial weight matrix W. According to the study by Kubara & Kopczewska (2023), the AIC of such a set of models is a non-linear function
with a clear minimum. This function searches among the models for the lowest AIC and the underlying structure of the neighbourhood -
this is the best W. The function reports two line plots: AIC and spatial parameter (rho or lambda), both depending on knn,
and returns the best spatial weight matrix W.

The function may run into computational problems on large data, which is typical of all spatial functions. The \code{sample_size} option allows
to set a smaller number of observations than in the original dataset to speed up the computation. If \code{sample_size} is equal to (or greater than)
the size of the dataset, all observations will be used.
}
\examples{
#To be done!!!

}
\references{
Kubara, M., & Kopczewska, K. (2023). Akaike information criterion in choosing the optimal k-nearest neighbours of the spatial weight matrix.
Spatial Economic Analysis, 1-19.
}
