% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootdbscan.R
\name{bootdbscan}
\alias{bootdbscan}
\alias{BOOTDBSCAN}
\title{Linijka nr 1 - function title}
\usage{
bootdbscan(data_sf, sample_size, times, eps, minPts)
}
\arguments{
\item{data_sf}{do opisu (obiekt sf lub data.frame - w data frame 1 kolumna musi być X coords, druga kolumna Y coords)}

\item{sample_size}{do opisu}

\item{times}{do opisu}

\item{eps}{do opisu}

\item{minPts}{do opisu}
}
\value{
\code{bootdbscan()} returns an object of class \code{bootdbscan} with the following components:

\item{cluster }{A integer vector with cluster assignments. Zero indicates noise points.}
\item{sample_size }{ value of the \code{sample_size} parameter.}
\item{times }{ value of the \code{times} parameter.}
\item{eps }{ value of the \code{eps} parameter.}
\item{minPts }{ value of the \code{minPts} parameter.}
}
\description{
Linijka nr 2 - description
}
\details{
Linijka nr 3 - details
}
\examples{
#To be done!!!

}
