% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootspatreg.R
\name{BootSpatReg}
\alias{BootSpatReg}
\title{Linijka nr 1 - function title}
\usage{
BootSpatReg(points_sf, iter, sample_size, eq, model_type, knn)
}
\arguments{
\item{points_sf}{do opisu (obiekt sf lub data.frame - w data frame 1 kolumna musi być X coords, druga kolumna Y coords).}

\item{iter}{Number of iterations}

\item{sample_size}{Sample size, must be less than or equal to the number of points in the dataset (\code{points_sf} parameter).
If \code{sample_size} is larger, it is automatically set to the number of points in the dataset.
We suggest that a value greater than 800 is not used for reasons of computational efficiency. (SPRAWDZIĆ)}

\item{eq}{an object of class \code{\link[stats:formula]{stats::formula()}} (or one that can be coerced to that class):
a symbolic description of the model to be used.}

\item{model_type}{one of: "SAR","SDM","SEM". Default set to "SDM"}

\item{knn}{chosen knn}
}
\value{
\code{BootSpatReg()} returns ... to be done.
}
\description{
Linijka nr 2 - description
}
\details{
Linijka nr 3 - details
}
\examples{
#To be done!!!

}
