% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta.R
\name{ETA}
\alias{ETA}
\alias{eta}
\title{Entropy and Tesselation for Agglomeration (ETA)}
\usage{
ETA(points_sf, region_sf, sample_size)

eta(points_sf, region_sf, sample_size)
}
\arguments{
\item{points_sf}{Object in \code{sf} or the \code{data.frame} class - in the case of a \code{data.frame} object, the first and second columns must contain X and Y coordinates.
For \code{data.frame}, make sure that the coordinates of the points are in the same coordinate system / projection as the \code{region_sf} object.}

\item{region_sf}{Polygon in the sf class that defines the boundary for points_sf.}

\item{sample_size}{Sample size, must be less than or equal to the number of points in the dataset (\code{points_sf} parameter).
If \code{sample_size} is equal to points_sf size, all points are used to construct Voronoi polygons.
If \code{sample_size} is less than the sample size, the functions will perform random sampling.
If \code{sample_size} is greater, it is automatically set to the number of points in the dataset.
The number of points affects the speed of Voronoi polygon computation.}
}
\value{
\code{ETA()} returns a list object and a plot. The list contains three elements:
\item{S_ent}{Empirical entropy of the tessellated point pattern.}
\item{H_ent}{ETA, degree of agglomeration.}
\item{n_points}{Number of points included in calculations.}
The plot illustrates the points used in the analysis and the boundary region,
Voronoi tesselation and displays values from the list.
}
\description{
For geolocated point patterns, the degree of agglomeration is calculated. It is expressed as the relative Shannon
entropy of the relative areas of tesselation tiles derived as Voronoi polygons. ETA (Entropy and Tesselation
for Agglomeration) takes values between 0 and 1. Values close to 1 reflect highly dispersed geolocated points
and spatially uniform distribution. The lower the value of ETA, the stronger the agglomeration.
}
\details{
The function calculates the relative areas of Voronoi polygons (ri) that sum to 1 such that sum(ri)=1.
The relative areas (ri) are an input to the Shannon entropy function H, where H=-sum(ri*log(ri)).
The maximum entropy Hmax is derived as Hmax=-1/n for n tiles. The relative entropy is calculated as H/Hmax
and interpreted as ETA.

In the case of spatially uniform distribution (regular location of points), ETA=1. The more points are
agglomerated, the lower the ETA. Points agglomerated in a single location should have ETA~0.

A number of points may be reduced at the tesselation stage if duplicate locations are identified in the dataset.
}
\examples{
# `sf` package required
library(sf)

# for random point pattern
eta.random<-st_sample(region_sf, 500, type="random")
eta.random<-st_as_sf(eta.random)
ETA(eta.random, region_sf, 500)

# for regular point pattern
eta.regular<-st_sample(region_sf, 500, type="regular")
eta.regular<-st_as_sf(eta.regular, crs=4326)
ETA(eta.regular, region_sf, 500)

# for clustered point pattern
eta.clust<-data.frame(X=rnorm(500, mean=22.8, sd=0.25), Y=rnorm(500, mean=51.2, sd=0.25))
plot(st_geometry(region_sf))
points(eta.clust)
ETA(eta.clust, region_sf, 500)

}
\references{
Kopczewska K., 2021, Entropy as measure of agglomeration. Interactions of business locations and housing transactions
in Warsaw metropolitan area. [In] Handbook on "Entropy, Complexity, and Spatial Dynamics: The Rebirth of Theory?",
Edward Elgar (editors: Aura Reggiani, Laurie Schintler, Danny Czamanski, Roberto Patuelli)
}
