% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fle.R
\name{FLE}
\alias{FLE}
\alias{fle}
\title{Linijka nr 1 - function title}
\usage{
FLE(points_sf, region_sf, nrows.raster = 50, ncols.raster = 50, w, r)
}
\arguments{
\item{points_sf}{do opisu (obiekt sf lub data.frame - w data frame 1 kolumna musi być X coords, druga kolumna Y coords). When using a simple data.frame, make sure that the coordinates of the points are in the same coordinate system / projection as the \code{region_sf} object.}

\item{region_sf}{do opisu (obiekt sf ale jako region)}

\item{nrows.raster}{raster row dimensions, default 50}

\item{ncols.raster}{raster column dimensions, default 50}

\item{w}{size of squared moving window (choose 3,5,7,9 etc), w and r parameters are mutually exclusive}

\item{r}{radius of circular ring, w and r parameters are mutually exclusive}
}
\value{
\code{FLE()} returns ... to be done.
}
\description{
Linijka nr 2 - description
}
\details{
Linijka nr 3 - details
}
\examples{
#To be done!!!

}
