% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastclust.R
\name{rastClustGWR}
\alias{rastClustGWR}
\title{Rasterisation of clustered Geographically Weighted Regression (GWR) coefficients}
\usage{
rastClustGWR(
  points_sf,
  eq,
  region_sf,
  nrows.raster = 50,
  ncols.raster = 50,
  nc,
  bw,
  adaptive = FALSE
)
}
\arguments{
\item{points_sf}{Geo-located points in \code{sf} or the \code{data.frame} class - in the case of a \code{data.frame} object, the first and second columns must contain X and Y coordinates.
For \code{data.frame}, make sure that the coordinates of the points are in the same coordinate system / projection as the \code{region_sf} object.
Data must be for a single period.}

\item{eq}{An object that defines the equation for the model, can be in the \code{\link[stats:formula]{stats::formula()}} class (or one that can be coerced to that class) or
a symbolic description of the model to be used.}

\item{region_sf}{Polygon in the \code{sf} class that defines the boundary for \code{points_sf}.}

\item{nrows.raster}{Raster row dimension, default 50.}

\item{ncols.raster}{Raster column dimension, default 50.}

\item{nc}{Number of clusters in k-means.}

\item{bw}{GWR model bandwidth, if missing, it is calculated using the \code{\link[GWmodel:bw.gwr]{GWmodel::bw.gwr()}} function (may be time-consuming); its value can be specified by the user.}

\item{adaptive}{Specification of the kernel in the GWR function, by default \code{adaptive=FALSE}, can be \code{TRUE} (used in case of highly uneven point density).}
}
\value{
\code{rastClustGWR()} returns the raster (with the number of rows and columns specified in the function call) that calculates the median of the k-means cluster ID -
it clusters the GWR coefficients. The function returns both the \code{terra} class object and its visualisation.
}
\description{
This is a multi-step algorithm for single period spatial point data. First, it estimates the GWR model.
Second, it uses k-means to cluster the model coefficients into k clusters. Third, it rasterises the data and computes
the median of the cluster IDs within the grid cell.
}
\details{
The \code{rastClustGWR()} function operates on the \code{\link[GWmodel:gwr.basic]{GWmodel::gwr.basic()}} function to compute Geographically Weighted Regression models,
\code{\link[stats:kmeans]{stats::kmeans()}} to cluster GWR coefficients, and \code{\link[terra:rasterize]{terra::rasterize()}} to count the medians of the cluster IDs in grid cells. All operations
are for single period data. The result, visualisation of grids coloured by cluster IDs, helps to detect heterogeneity in the relationships
modelled with GWR. Rasterisation helps to specify a common spatial reference for point data.

This algorithm is extended in \code{\link[=STS]{STS()}} function, which tests the spatio-temporal stability of clusters of GWR coefficients
(developed by Kopczewska & Ćwiakowski, 2021).
}
\examples{
# The form of the equation to be estimated:
# Companies' return on assets (ROA) depends on their size, sector and relative location.
eq<-roa~empl+dummy.prod+dummy.constr+dummy.serv+dist.big.city

firms2015.sf<-firms_sf[firms_sf$year==2015,]

# It takes a while to estimate the GWR model
rcGWR<-rastClustGWR(firms2015.sf, eq, region_sf, nc=5, bw=0.05)
rcGWR

}
\references{
Kopczewska, K., & Ćwiakowski, P. (2021). Spatio-temporal stability of housing submarkets. Tracking spatial location of clusters of
geographically weighted regression estimates of price determinants. Land Use Policy, 103, 105292.
https://www.sciencedirect.com/science/article/pii/S0264837721000156
}
