% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benford.R
\name{SpatBenfordTest}
\alias{SpatBenfordTest}
\alias{spatbenfordtest}
\title{Benford's law spatial point pattern conformity test}
\usage{
SpatBenfordTest(data_sf, sample_size, var_name = NULL)

spatbenfordtest(data_sf, sample_size, var_name = NULL)
}
\arguments{
\item{data_sf}{Object in \code{sf} or the \code{data.frame} class - in the case of a \code{data.frame} object, the first and second columns must contain X and Y coordinates.}

\item{sample_size}{Number of points to be used in the analysis. It must be less than or equal to the number of points in the dataset.
Huge datasets can cause computational problems when testing for Benford's law conformity due to the nxn size matrix of mutual distances.}

\item{var_name}{Column name with additional numeric variable. If empty, a 2d distribution of Euclidean distances is tested, if given, a 3d distribution is tested.}
}
\value{
\code{SpatBenfordTest()} returns the test for Benford's law conformity.
}
\description{
Spatial point patterns can conform to Benford's law if the mutual distances between points conform to Benford's law.
This occurs in the case of a mixture of regular, random and clustered point patterns composed in stable proportions.
The function calculates the mutual Euclidean distances between all points in the dataset and tests whether
this distribution conforms to Benford's law.
}
\details{
The function returns the Benford's law conformity test of geolocated spatial points using the \code{benford()} function
from the \code{benford.analysis} package. Euclidean distances between points are calculated using the \code{dist()} function.
}
\examples{
#To be done!!!

}
