% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_spatialWarsaw-package.R
\docType{package}
\name{spatialWarsaw-package}
\alias{spatialWarsaw}
\alias{spatialWarsaw-package}
\title{spatialWarsaw: Spatial analysis on point data including clustering measures and bootstrap methods}
\description{
A collection of tools for spatial analysis on point data including clustering measures and the ability to bootstrap large datasets.
(Use four spaces when indenting paragraphs within the Description.)
}
\references{
To be done
}
\author{
Katarzyna Kopczewska and Mateusz Kopyt
}
