% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssr.R
\name{ssr}
\alias{ssr}
\title{ssr() function}
\usage{
ssr(data_sf, clusters, eq, family = binomial, ...)
}
\arguments{
\item{data_sf}{Geo-located points in \code{sf} or the \code{data.frame} class - in the case of a \code{data.frame} object, the first and second columns must contain X and Y coordinates.
\emph{\strong{NOTE! The same \code{data_sf} object must also be used to obtain a cluster split using the \code{\link[=ClustConti]{ClustConti()}} or \code{\link[=ClustDisjoint]{ClustDisjoint()}} function.}}}

\item{clusters}{Object of class \code{clust} as output of the function \code{\link[=ClustConti]{ClustConti()}} or \code{\link[=ClustDisjoint]{ClustDisjoint()}} containing a vector of assignments of observations from the object \code{data_sf}
to individual clusters and the name of the function used for the division.}

\item{eq}{an object of class \code{\link[stats:formula]{stats::formula()}} (or one that can be coerced to that class):
a symbolic description of the model to be used.}

\item{family}{A \code{family} object  consistent with \code{\link[stats:family]{stats::family()}}: a description of the error distribution
and link function to be used in the model. The default value is \code{binomial} with logit link function.}

\item{...}{Other parameters passed to \code{\link[stats:glm]{stats::glm()}} during the calculation.}
}
\value{
Will be updated soon
}
\description{
Will be updated soon
}
\details{
Will be updated soon
}
\examples{
#To be done!!!

}
