% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssr.R
\name{ssr}
\alias{ssr}
\title{Spatial Switching Regimes (SSR): estimation and comparison of econometric models in groups of different densities}
\usage{
ssr(data_sf, clusters, eq, family = binomial, ...)
}
\arguments{
\item{data_sf}{Geo-located points in \code{sf} or the \code{data.frame} class - in the case of a \code{data.frame} object, the first and second columns must contain X and Y coordinates.
\emph{\strong{NOTE! The same \code{data_sf} object must also be used to obtain a cluster split using the \code{\link[=ClustConti]{ClustConti()}} or \code{\link[=ClustDisjoint]{ClustDisjoint()}} function.}}}

\item{clusters}{Object of class \code{clust} as output of the function \code{\link[=ClustConti]{ClustConti()}} or \code{\link[=ClustDisjoint]{ClustDisjoint()}} containing a vector of assignments of observations from the object \code{data_sf}
to individual clusters and the name of the function used for the division.}

\item{eq}{An object that defines the equation for the model, can be in the \code{\link[stats:formula]{stats::formula()}} class (or one that can be coerced to that class) or a symbolic description of the model to be used.}

\item{family}{A \code{family} object  consistent with \code{\link[stats:family]{stats::family()}}: a description of the error distribution
and link function to be used in the model. The default value is \code{binomial} with logit link function.}

\item{...}{Other parameters passed to \code{\link[stats:glm]{stats::glm()}} during the calculation.}
}
\value{
\code{ssr()} returns the table with a summary of models estimated in density groups and a list object:
\item{data}{provides data by density groups.}
\item{coords}{provides all xy coordinates of point data, without the density group subdivision information.}
\item{cluster}{Provides a clustering vector that assigns all point data to density clusters.}
\item{type}{provides the type of clustering based on the \code{clusters} argument, one of "ClustDisjoint" or "ClustConti".}
\item{models}{privides the basic models (without significance tests) in each density cluster.}
The output also includes a \code{texreg} table which summarises the models in density groups and the size of density clusters.
}
\description{
The \code{ssr()} function works on geolocated point data. It uses input from the \code{\link[=ClustConti]{ClustConti()}} or \code{\link[=ClustDisjoint]{ClustDisjoint()}} functions from the \code{spatialWarsaw} package to divide points
into density groups - within these subsamples it estimates the econometric model given by the equation. An important point is that the size and order of the datasets used to obtain density groups and
the econometric model must be the same.
}
\details{
Function returns a table with a summary of the models estimated in density groups.
}
\examples{
# The following example is on selected data from the firms_sf dataset included in the package
sub_sf<-firms_sf[1:5000,]
# One to choose
# clust<-ClustConti(sub_sf, clusters=4, noise =0.2, minPts0=30)
clust<-ClustDisjoint(sub_sf, noise=c(0.8, 0.6, 0.4, 0.2), minPts=30)

# Density cluster plot
sub_sf$cluster<-clust$cluster
plot(sub_sf["cluster"], key.pos=4, pch=".", cex=2, main="Density clusters")

# SSR models for the binary dependent variable
# The form of the equation to be estimated:
eq<-dummy.prod~empl+roa+dist.big.city

my.ssr<-ssr(sub_sf, clust, eq, family=binomial)
my.ssr

# Output with plot using ssrShowModels() function.
ssrShowModels(my.ssr, plot=TRUE)

}
\seealso{
\code{\link[=ssrShowModels]{ssrShowModels()}}
}
