% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssr.R
\name{ssr}
\alias{ssr}
\alias{show_models}
\title{Linijka nr 1 - function title}
\usage{
ssr(
  formula,
  data.spsf,
  type,
  clusters,
  minPts = 5,
  r_p = 0.001,
  sep,
  bootstrap = FALSE,
  sample_size,
  times,
  family = stats::binomial(link = "logit"),
  eps_r = 1e-15,
  eps_np = 0.01,
  ...
)

show_models(ssr_model, plot = TRUE, ...)
}
\arguments{
\item{formula}{do opisu}

\item{data.spsf}{do opisu}

\item{type}{do opisu}

\item{clusters}{do opisu}

\item{minPts}{do opisu}

\item{r_p}{do opisu}

\item{sep}{do opisu}

\item{bootstrap}{do opisu}

\item{sample_size}{do opisu}

\item{times}{do opisu}

\item{family}{do opisu}

\item{eps_r}{do opisu}

\item{eps_np}{do opisu}

\item{...}{other paramethers (do opisu)}

\item{ssr_model}{do opisu}

\item{plot}{do opisu}
}
\description{
Linijka nr 2 - description (razem do funkcji ssr i show_models)
}
\details{
Linijka nr 3 - details (razem do funkcji ssr i show_models)
}
\examples{
#To be done!!!
}
