% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssr.R
\name{ssrShowModels}
\alias{ssrShowModels}
\title{Linijka nr 1 - ssrShowModels() title}
\usage{
ssrShowModels(ssr_model, plot = TRUE, ...)
}
\arguments{
\item{ssr_model}{do opisu}

\item{plot}{do opisu}
}
\value{
ssrShowModels description (DZIAŁA)
}
\description{
A short description of ssrShowModels
}
\details{
Linijka nr 3 - details (razem do funkcji ssr i show_models?)
}
\examples{
#To be done!

}
