% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssr.R
\name{ssrShowModels}
\alias{ssrShowModels}
\title{Friendly summary and plot for ssr objects (ssr() function result)}
\usage{
ssrShowModels(ssr_model, plot = TRUE, ...)
}
\arguments{
\item{ssr_model}{Result object of the \code{\link[=ssr]{ssr()}} function from this package. The class of \code{ssr_model} object must only be 'ssr'.}

\item{plot}{Logical; indicates whether the function should generate a plot, default \code{plot=TRUE}.}

\item{...}{Other parameters passed to \code{\link[stargazer:stargazer]{stargazer::stargazer()}}.}
}
\value{
This function returns the summary table of econometric models in density clusters. If \code{plot=TRUE} it will also return
the the point plot with division into clusters.
}
\description{
The \code{ssrShowModels()} function with the \code{plot=TRUE} option shows the summary of the models and plots the points divided
by the applied density clusters.
}
\details{
The plot illustrates the division of geolocated points into density clusters. This partitioning is used in
the \code{\link[=ssr]{ssr()}} function for subgroup model estimation.
}
\examples{
#To be done!

}
