% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssr.R
\name{ssrShowModels}
\alias{ssrShowModels}
\title{ssrShowModels() function}
\usage{
ssrShowModels(ssr_model, plot = TRUE, ...)
}
\arguments{
\item{ssr_model}{Result object of the \code{\link[=ssr]{ssr()}} function from this package. The class of \code{ssr_model} object must only be 'ssr'.}

\item{plot}{Logical; indicates whether the function should generate a plot, default \code{plot=TRUE}.}

\item{...}{Other parameters passed to \code{\link[stargazer:stargazer]{stargazer::stargazer()}}.}
}
\value{
Will be updated soon
}
\description{
Will be updated soon
}
\details{
Will be updated soon
}
\examples{
#To be done!

}
